+++ lib/Backend/analytics.dart
@@ -5,6 +5,7 @@ import 'package:flutter/foundation.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:synchronized/synchronized.dart';
 import 'package:tail_app/Backend/dynamic_config.dart';
+import 'package:tail_app/Backend/utilities/settings.dart';
 import 'package:tail_app/Backend/wear_bridge.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -119,11 +120,7 @@ Future<Map<String, String>> _getSettingsProps({
     selectedLocale,
     defaultValue: "Not Set",
   ).toString();
-  props['Developer Mode'] = HiveProxy.getOrDefault(
-    settings,
-    showDebugging,
-    defaultValue: showDebuggingDefault,
-  ).toString();
+  props['Developer Mode'] = isDeveloperEnabled.toString();
   props['Custom App Color'] =
       (HiveProxy.getOrDefault(
                 settings,